backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
supabase==2.5.0
openai==1.30.1
pdfplumber==0.11.1
python-docx==1.1.2
python-dotenv==1.0.1
gunicorn==22.0.0
PyJWT==2.8.0
python-pptx==1.0.2
